+++ web-components/src/localization/index.ts
@@ -1,13 +1,7 @@
 import { configureLocalization } from "@lit/localize"
 import { sourceLocale, targetLocales } from "./dist/locale-codes"
-import { delay } from "../utils"
+import { languageCode } from "../signals/app"
 
-/**
- * Get the browser locale, it will keep only the language part
- */
-export const getBrowserLocale = () => {
-  return (navigator.language || navigator.languages[0]).split("-")[0]
-}
 /**
  * Configure the localization, it will load the locale files and set the source and target locales
  */
@@ -17,19 +11,11 @@ export const { getLocale, setLocale } = configureLocalization({
   loadLocale: (locale: string) => import(`./localization/locales/${locale}.js`),
 })
 
-// Wait for the locale to be set
-export const isLocaleSet = false
-
-/**
- * Get the locale if it's set or delay 100ms to get it
- */
-export const getLocaleAfterInit = async (): Promise<string> => {
-  let index = 0
-
-  // Delay to wait for the locale to be set - 1s max
-  while (!isLocaleSet && index < 10) {
-    await delay(100)
-    index++
+export const setLanguageCode = (newLanguageCode: string) => {
+  if (!targetLocales.includes(newLanguageCode as any)) {
+    throw new Error(`Invalid language code: ${newLanguageCode}`)
   }
-  return getLocale()
+  // Set the language code (en, fr, ..)
+  setLocale(newLanguageCode)
+  languageCode.set(newLanguageCode)
 }